Slot manager: Refactor to tsx
